+++ xenon-common/src/main/java/com/vmware/xenon/common/http/netty/NettyWebSocketRequestHandler.java
@@ -169,7 +169,8 @@ public class NettyWebSocketRequestHandler extends SimpleChannelInboundHandler<Ob
             if (token == null) {
                 String cookie = nettyRequest.headers().get(HttpHeaderNames.COOKIE);
                 if (cookie != null) {
-                    token = CookieJar.decodeCookies(cookie).get(AuthenticationConstants.XENON_JWT_COOKIE);
+                    token = CookieJar.decodeCookies(cookie)
+                            .get(AuthenticationConstants.REQUEST_AUTH_TOKEN_COOKIE);
                 }
 
             }